fix: change `streams` to be a SetNestedAttribute
This commit modifies the connection resource's `streams` attribute to be
a `SetNestedAttribute` instead of a List. This is so that we can
properly compute plan's that treat re-ordering of elements (streams) as
a noop (if no attributes of the stream have been changed).

Previously if you added a new stream to a connection resource in the
middle of the list, or changed the order, the provider would compute
that as a change that had to be applied.

Additionally, we use a customer SetPlanModifier that may be
parameterized with a `key` which is used to uniquely identify the
elements in the set.

NOTE: the current implementation of the plan modifier `uniqueByKey`
assumes that the elements of the set are Objects. Since this is the only
use case for this modifier, this seems reasonable for now.